[WIP] fix tests fatturapa in
